frontend: Refactor Customer Page to use SubNav and subroutes
