+++ clients/apps/web/src/app/(main)/dashboard/[organization]/(header)/customers/[customerId]/members/page.tsx
@@ -1,16 +1,17 @@
-import { CustomerOverview } from '@/components/Customer/CustomerPage/CustomerOverview'
+import { CustomerMembers } from '@/components/Customer/CustomerPage/CustomerMembers'
 import { CustomerPageShell } from '@/components/Customer/CustomerPage/CustomerPageShell'
 import { Metadata } from 'next'
+import { notFound } from 'next/navigation'
 import {
   CustomerPageParams,
   generateCustomerMetadata,
   getOrganizationAndCustomerOrNotFound,
-} from './getCustomer'
+} from '../getCustomer'
 
 export async function generateMetadata(props: {
   params: Promise<CustomerPageParams>
 }): Promise<Metadata> {
-  return generateCustomerMetadata(await props.params)
+  return generateCustomerMetadata(await props.params, 'Members')
 }
 
 export default async function Page(props: {
@@ -20,13 +21,20 @@ export default async function Page(props: {
   const { organization, customer } =
     await getOrganizationAndCustomerOrNotFound(params)
 
+  if (
+    !organization.feature_settings?.member_model_enabled ||
+    customer.type !== 'team'
+  ) {
+    notFound()
+  }
+
   return (
     <CustomerPageShell
       key={customer.id}
       organization={organization}
       customer={customer}
     >
-      <CustomerOverview organization={organization} customer={customer} />
+      <CustomerMembers organization={organization} customer={customer} />
     </CustomerPageShell>
   )
 }
