+++ clients/apps/web/src/components/Customer/CustomerPage/CustomerTrendStatBox.tsx
